代码整理

+++ hardware/cortexM3/STM32F103/CMIOT3V1/port/wind_type.h
@@ -33,7 +33,7 @@ typedef unsigned char w_uint8_t;
 #endif
 
 #ifndef w_uint16_t
-typedef unsigned int w_uint16_t;
+typedef unsigned short w_uint16_t;
 #endif
 
 #ifndef w_uint32_t
@@ -45,19 +45,19 @@ typedef unsigned long long w_uint64_t;
 #endif
 
 #ifndef w_int8_t
-typedef char w_int8_t; 
+typedef signed char w_int8_t; 
 #endif
 
 #ifndef w_int16_t
-typedef int w_int16_t; 
+typedef signed short w_int16_t; 
 #endif
 
 #ifndef w_int32_t
-typedef long w_int32_t; 
+typedef signed long w_int32_t; 
 #endif
 
 #ifndef w_int64_t
-typedef long long w_int64_t; 
+typedef signed long long w_int64_t; 
 #endif
